Move the request logic to the from the map to the client
